Explicit conversions for msvc

+++ crypto_kem/sntrup761/clean/crypto_core_inv3sntrup761.c
@@ -88,16 +88,16 @@ int PQCLEAN_SNTRUP761_CLEAN_crypto_core_inv3sntrup761(unsigned char *outbytes, c
         }
 
         for (i = 0; i < p + 1; ++i) {
-            g[i] = F3_freeze((int16) g[i] + sign * f[i]);
+            g[i] = F3_freeze((int16) (g[i] + sign * f[i]));
         }
         for (i = 0; i < p + 1; ++i) {
-            r[i] = F3_freeze((int16) r[i] + sign * v[i]);
+            r[i] = F3_freeze((int16) (r[i] + sign * v[i]));
         }
 
         for (i = 0; i < p; ++i) {
             g[i] = g[i + 1];
         }
-        g[p] = 0;
+        g[p] = (int16) 0;
     }
 
     sign = (int) f[0];
